requirements.txt
redis==3.5.3
python_socketio==5.2.0
aioredis==1.3.1